+++ tests/UC_3_plan_meals_for_week/End_to_end_tests/test_UC3_E2E_1.py
@@ -1,0 +1,6 @@
+
+
+def UC3_E2E_1(test_client):
+    response = test_client.get("/", follow_redirects=True)
+    assert response.status_code == 200
+
